requirements.txt
Django==4.2
djangorestframework==3.14
dicttoxml==1.7.5
drf-spectacular==0.27.2
drf-spectacular-sidecar==2024.7.1
gunicorn>=21.2.0